+++ depchain/src/main/java/pt/tecnico/ulisboa/protocol/TransferFromReq.java
@@ -1,5 +1,7 @@
 package pt.tecnico.ulisboa.protocol;
 
+import java.math.BigInteger;
+
 // Transfer From request (e.g., for approved spending)
 public class TransferFromReq extends ClientReq {
     // TODO: These should all be addresses
@@ -7,9 +9,9 @@ public class TransferFromReq extends ClientReq {
     private String from; // transfer from this guy
     private String to; // transfer to this guy
 
-    private double amount; // amount to transfer
+    private BigInteger amount; // amount to transfer
 
-    public TransferFromReq(Integer id, long count, String spender, String from, String to, double amount) {
+    public TransferFromReq(Integer id, long count, String spender, String from, String to, BigInteger amount) {
         super(id, count);
         this.spender = spender;
         this.from = from;
@@ -34,12 +36,13 @@ public class TransferFromReq extends ClientReq {
         return to;
     }
 
-    public double getAmount() {
+    public BigInteger getAmount() {
         return amount;
     }
 
     @Override
     public String toString() {
         return String.format("TransferFromReq(id=%d, spender=%s, from=%s, to=%s, amount=%.2f, count=%d)",
-                id, spender, from, to, amount, count);}
+                id, spender, from, to, amount, count);
+    }
 }
